moves builder implementation validation to cmd
